sm-909 int now accepts double and negative

+++ lib/src/answer_format/integer_answer_format.dart
@@ -7,7 +7,7 @@ part 'integer_answer_format.g.dart';
 
 @JsonSerializable()
 class IntegerAnswerFormat implements AnswerFormat {
-  final int? defaultValue;
+  final double? defaultValue;
   final String hint;
   final IntegerQuestionResult? savedResult;
   final bool isChildQuestion;
